FlightBooking/requirements.txt
asgiref==3.5.1
automium==0.2.6
automium-web==0.1.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
defusedxml==0.7.1
Django==4.0.4
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
Flask==2.1.2
httpie==3.2.1
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
legacy==0.1.6
Markdown==3.3.6
MarkupSafe==2.1.1
multidict==6.0.2
oauthlib==3.2.0
pycparser==2.21
Pygments==2.12.0
PyJWT==1.7.1
PySocks==1.7.1
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rich==12.4.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
views==0.3
Werkzeug==2.1.2
zipp==3.8.0
